# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для управления студентами, классами и лекциями",
        "version": "1.0"
    },
    "paths": {
...
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Получить список всех студентов",
                "operationId": "ea63bf4dd71d5b07e457e9c1f86a5f48",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Создать студента",
                "operationId": "517b42e732f558e5f11885d11ead9be5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Студент создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Получить информацию о конкретном студенте",
                "operationId": "38806f18ebd0743308833c09a0c201f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Обновить студента",
                "operationId": "db547712a1caf50ee1a525d93d9d9181",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Студент обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Удалить студента",
                "operationId": "3e1c01a7953da510160724c5f6ba7a8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Студент удален"
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreStudentRequest": {
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "student_class_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateStudentRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "student_class_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Student": {
                "title": "Student",
                "description": "Модель студента",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "student_class_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}